Dev new

+++ js/productos.js
@@ -12,6 +12,9 @@ const { createApp } = Vue
         imagen:"",
         telefono:0,
         Dni:0,
+        direccion:"",
+        fechnac: new Date(),
+        feching: new Date(),
     }  
     },
     methods: {
@@ -28,6 +31,9 @@ const { createApp } = Vue
                 })
         },
         eliminar(producto) {
+            if (!confirm('¿Estás seguro/a de que deseas eliminar este producto?')) {
+                return; 
+            }
             const url = this.url+'/' + producto;
             var options = {
                 method: 'DELETE',
@@ -39,10 +45,38 @@ const { createApp } = Vue
                 })
         },
         grabar(){
+            if (this.nombre.trim() === '') {
+                alert('El campo "nombre" es obligatorio');
+            return; 
+            }
+            if (!Number.isInteger(this.Dni) || this.Dni <= 0) {
+                alert('El campo "Dni" debe ser un número válido');
+            return;
+            }
+            if (!Number.isInteger(this.telefono) || this.telefono <= 0) {
+                alert('El campo "telefono" debe ser un número válido');
+            return;
+            }
+            if (this.direccion.trim() === '') {
+                alert('El campo "direccion" es obligatorio');
+            return;
+            }
+            const fechaActual = new Date();
+            if (this.fechnac >= fechaActual) {
+                alert('Debe Ingresar la fecha de Nacimiento');
+            return;
+            }
+            if (this.feching <= this.fechnac) {
+                alert('Debe Ingresar la fecha de Ingreso a la empresa');
+            return;
+            }
             let producto = {
                 nombre:this.nombre,
                 Dni: this.Dni,
                 telefono: this.telefono,
+                direccion:this.direccion,
+                fechnac:this.fechnac,
+                feching:this.feching,
                 imagen:this.imagen
             }
             var options = {
